Added construction table

+++ src/main/java/net/threetag/threecore/base/block/StirlingGeneratorBlock.java
@@ -45,7 +45,7 @@ public class StirlingGeneratorBlock extends MachineBlock {
     public void animateTick(BlockState stateIn, World worldIn, BlockPos pos, Random rand) {
         if (stateIn.get(LIT)) {
             double d0 = (double) pos.getX() + 0.5D;
-            double d1 = (double) pos.getY();
+            double d1 = pos.getY();
             double d2 = (double) pos.getZ() + 0.5D;
 
             if (rand.nextDouble() < 0.1D) {
